HH-218702 move ExecuteOnDataSource validation to nab-jdbc module
