+++ nab-jdbc/src/main/java/ru/hh/nab/jdbc/validation/DataSourcesReadyTarget.java
@@ -1,4 +1,4 @@
-package ru.hh.nab.hibernate.transaction;
+package ru.hh.nab.jdbc.validation;
 
 import java.util.List;
 import javax.sql.DataSource;
